+++ themes/teabag/views/footer.php
@@ -1,0 +1,10 @@
+<div class="container footer-margin">
+	<div class="row clearfix"><div class="grid_12"><hr class="styled" /></div></div>
+</div>
+<div class="container">
+	<div class="row clearfix">
+		<footer class="grid_12">
+			<p class="footer-copy">Bancha - Generated in {elapsed_time} seconds. - <?php languages(); ?></p>
+		</footer>
+	</div>
+</div>