+++ themes/teabag/views/templates/default.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Default website (generic page) template
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license		GNU/GPL (General Public License)
+ * @link		http://squallstar.it
+ *
+ */
+
+render('header');
+
+content_render();
+
+render('footer');