+++ themes/teabag/views/templates/feed.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Feed template (Core)
+ *
+ * Tipically you should not touch this file.
+ * Feel free to extend it in your theme
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license		GNU/GPL (General Public License)
+ * @link		http://squallstar.it
+ *
+ */
+
+echo $feed;
+
+?>