+++ themes/teabag/views/type_templates/Blog/detail.php
@@ -1,0 +1,78 @@
+<?php
+/**
+ * Blog Detail View
+ *
+ * Content detail - Sandbox Theme
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license		GNU/GPL (General Public License)
+ *
+ */
+
+//Here you can save a new comment when a form is posted
+if ($this->input->post('message') && $this->input->post('author'))
+{
+	//Check the sandbox/views/extra/comments-save.php file
+	render('extra/comments-save');	
+}
+
+//Finally we get the comments linked to the current post that we are viewing
+$comments =& $record->related('comments');
+
+?>
+
+<div class="container">
+	<div class="row clearfix">
+		<div class="grid_12">
+			<article>
+				<h1><?php echo page('title'); ?> &rarr; <?php echo record('title'); ?></h1>
+				<h3>Postd on <?php echo date('d/m/Y', record('date_insert')); ?></h3>
+
+				<p><?php echo record('content'); ?></p>
+			</article>
+		</div>
+	</div>
+</div>
+
+<div class="container">
+	<div class="row clearfix">
+		<div class="grid_12">
+			<h3><?php echo count($comments); ?> Comments</h3>
+			<?php
+			//Here we display the comments
+			if (is_array($comments) && count($comments)) {
+				echo '<ul>';
+				foreach ($comments as $comment) {
+					echo '<li>' . date('d/m/Y H:i', $comment->get('date_insert'))
+					   . ' by <strong>' . $comment->get('author') . '</strong>'
+					   . '<p>' . strip_tags($this->security->xss_clean($comment->get('content'))) . '</p></li>';
+				}
+				echo '</ul>';
+			}
+			?>
+
+			<hr />
+			<h3>Add a new comment</h3>
+			<form action="<?php echo current_url(); ?>" method="post">
+				<fieldset>
+					<label for="author"><?php echo _('Name'); ?></label><br />
+					<input type="text" id="author" name="author" /><br /><br />
+
+					<label for="email"><?php echo _('E-mail address'); ?></label><br />
+					<input type="text" id="email" name="email" /><br /><br />
+
+					<label for="message"><?php echo _('Message'); ?></label><br />
+					<textarea id="name" name="message"></textarea><br /><br />
+
+					<input type="submit" value="<?php echo _('Send'); ?>" />
+				</fieldset>
+			</form>
+		</div>
+	</div>
+</div>
+
+<?php
+/* End of file detail.php */
+/* Location: /type_templates/Blog/detail.php */