+++ themes/teabag/views/type_templates/Default-Content/detail.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Detail View
+ *
+ * Content detail - Sandbox Theme
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license		GNU/GPL (General Public License)
+ *
+ */
+
+?>
+
+<div class="container">
+	<div class="row clearfix">
+		<div class="grid_12">
+			<h1><?php echo page('title'); ?></h1>
+
+			<h2><?php echo record('title'); ?></h2>
+			<p><?php echo record('content'); ?></p>
+		</div>
+	</div>
+</div>
+<?php
+/* End of file detail.php */
+/* Location: /type_templates/Default-Content/detail.php */